pyproject.toml
[tool.poetry]
name = "zalupa-discord-bot"
version = "0.1.0"
description = ""
authors = ["troyan-dy <troyan-dy@sber-zvuk.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "3.9.9"
"discord.py" = "^1.7.3"
certifi = "^2021.10.8"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
